♻️ refactor: separate context in card component
