+++ packages/card/src/components/CardBody.tsx
@@ -2,7 +2,7 @@
 import { jsx } from '@emotion/react';
 import { useContext } from 'react';
 import { createClassVariant } from '@jdesignlab/theme';
-import { CardContext } from './CardContext';
+import { CardContext } from '../context';
 import createContentStyle from '../styles/createContentStytle';
 import { DEFAULT_BORDER_COLOR } from '../constants';
 import type { CardBodyProps } from '../types';
