+++ packages/card/src/components/CardFooter.tsx
@@ -3,7 +3,7 @@ import { jsx } from '@emotion/react';
 import { useContext } from 'react';
 import { createClassVariant } from '@jdesignlab/theme';
 import { DEFAULT_BORDER_COLOR } from '../constants';
-import { CardContext } from './CardContext';
+import { CardContext } from '../context';
 import createContentStyle from '../styles/createContentStytle';
 import type { CardFooterProps } from '../types';
