+++ packages/card/src/components/CardHeader.tsx
@@ -2,7 +2,7 @@
 import { css, jsx } from '@emotion/react';
 import { useContext } from 'react';
 import { createClassVariant } from '@jdesignlab/theme';
-import { CardContext } from './CardContext';
+import { CardContext } from '../context';
 import createContentStyle from '../styles/createContentStytle';
 import type { CardHeaderProps } from '../types';
 import { DEFAULT_BORDER_COLOR } from '../constants';
